% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divModule.R
\name{divModule}
\alias{divModule}
\title{Dividing Modules.}
\usage{
divModule(
  case_union,
  control_union,
  deepSplit = 4,
  minModuleSize = 20,
  soft_pow = 5
)
}
\arguments{
\item{case_union}{a matrix of integrated diseased network.}

\item{control_union}{a matrix of integrated healthy network.}

\item{deepSplit}{numerical. For method "hybrid", can be either logical or integer in the range 0 to 4. For method "tree", must be logical. In both cases, provides a rough control over sensitivity to cluster splitting. The higher the value (or if TRUE), the more and smaller clusters will be produced. For the "hybrid" method, a finer control can be achieved via maxCoreScatter and minGap below.}

\item{minModuleSize}{numerical. Minimum module size. Set to 20 by default.}

\item{soft_pow}{numerical. Soft power for module division. Set to 5 by default.}
}
\value{
divided modules.
}
\description{
This is a function to divide integrated networks into different modules.
}
\examples{
data(testData)
netAll = getNetwork(case_dir = mydata[[1]],
     control_dir = mydata[[2]],
     net_case_dir = mydata[[3]],
     net_control_dir = mydata[[4]])
modAll = divModule(netAll[[1]], netAll[[2]])

}
