% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetwork.R
\name{getNetwork}
\alias{getNetwork}
\title{Integrating Networks.}
\usage{
getNetwork(case_dir, control_dir, net_case_dir, net_control_dir)
}
\arguments{
\item{case_dir}{string.The directory of diseased data set or a single diseased file.}

\item{control_dir}{string.The directory of healthy data set or a single healthy file.}

\item{net_case_dir}{string.The directory of network correlation of diseased data set or a single diseased network file.}

\item{net_control_dir}{string.The directory of network correlation of healthy data set or a single healthy network file.}
}
\value{
integrated diseased and healthy networks.
}
\description{
This is a function to integrate data set from different studies based on network.
}
\examples{
data(testData)
netAll = getNetwork(case_dir = mydata[[1]],
     control_dir = mydata[[2]],
     net_case_dir = mydata[[3]],
     net_control_dir = mydata[[4]])

}
