% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netPlot.R
\name{netPlot}
\alias{netPlot}
\title{Network plot.}
\usage{
netPlot(
  result,
  num.top = 5,
  num.score = 30,
  e.th = 0.4,
  nodeSize = 2,
  nodeDize = 1.2,
  edgeSize = 0.2,
  edgeDize = 0.3,
  arrowSize = 0,
  my.layout = layout_as_star,
  my.label = TRUE
)
}
\arguments{
\item{result}{the result list of NetMoss.}

\item{num.top}{numeric. The number of top taxon to be highlighted in the paired networks.}

\item{num.score}{numeric. The number of taxon to be plotted in the NetMoss score barplot and point plot.}

\item{e.th}{numeric. The threshold of microbial correlations in the networks. Edges greater than this threshold are shown in the networks.}

\item{nodeSize}{numeric. The size of nodes in the networks.}

\item{nodeDize}{numeric. The size of nodes in the networks. The larger the value, the greater the difference in the size of the nodes.}

\item{edgeSize}{numeric. The size of edges in the networks.}

\item{edgeDize}{numeric. The size of edges in the networks. The larger the value, the greater the difference in the size of the edges}

\item{arrowSize}{numeric. The size of arrow in the networks.}

\item{my.layout}{the layout of network.}

\item{my.label}{logic. If TRUE then the label of the nodes will be plotted.}
}
\value{
case and control networks
}
\description{
This is a function to plot combined networks.
}
\examples{
data(testData)
nodes_result = NetMoss(case_dir = mydata[[1]],
     control_dir = mydata[[2]],
     net_case_dir = mydata[[3]],
     net_control_dir = mydata[[4]])
netPlot(nodes_result)


}
