% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/netROC.R
\name{netROC}
\alias{netROC}
\title{Plot the combined ROC}
\usage{
netROC(
  case_dir,
  control_dir,
  marker,
  metadata,
  plot.roc = TRUE,
  train.num = 20
)
}
\arguments{
\item{case_dir}{string.The directory of diseased data set or a single diseased file.}

\item{control_dir}{string.The directory of healthy data set or a single healthy file.}

\item{marker}{the combined markers identified by NetMoss.}

\item{metadata}{the metadata of all input studies.}

\item{plot.roc}{logical.If TURE then the combined ROC will be plotted.}

\item{train.num}{numerical. training times for the module. 20 was set by default.}
}
\value{
a TPR and FPR table of the combined study.
}
\description{
This is a function to plot ROC of combined study.
}
\examples{
data(testData)
case_dir = mydata[[1]]
control_dir = mydata[[2]]
nodes_result = NetMoss(case_dir = case_dir,
     control_dir = control_dir,
     net_case_dir = mydata[[3]],
     net_control_dir = mydata[[4]])
result = nodes_result[[1]]
marker = data.frame(result[which(result$p.adj < 0.05),])
marker = data.frame(marker[which(marker$NetMoss_Score > 0.3),])
rownames(marker) = marker$taxon_names
metadata = mydata[[5]]
myROC = netROC(case_dir = case_dir,
     control_dir = control_dir,
     marker = marker,
     metadata = metadata,
     plot.roc = TRUE,
     train.num = 20)

}
