% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NetzGo.R
\name{NetzGO}
\alias{NetzGO}
\title{Calculating NetMoss Score}
\usage{
NetzGO(
  control_mat,
  case_mat,
  control_dist,
  case_dist,
  control_mod,
  case_mod,
  scaled = TRUE
)
}
\arguments{
\item{control_mat}{a matrix of integrated healthy network.}

\item{case_mat}{a matrix of integrated diseased network.}

\item{control_dist}{a distance matrix among healthy models.}

\item{case_dist}{a distance matrix among diseased models.}

\item{control_mod}{module division of each microbe in healthy network.}

\item{case_mod}{module division of each microbe diseased network}

\item{scaled}{logical.If TURE then the NetMoss score are scaled.}
}
\value{
NetMoss score and module division results of each bacterium.
}
\description{
This is a function to calculate NetMoss score for each bacterium.
}
\examples{
data(testData)
netAll = getNetwork(case_dir = mydata[[1]],
     control_dir = mydata[[2]],
     net_case_dir = mydata[[3]],
     net_control_dir = mydata[[4]])
modAll = divModule(netAll[[1]], netAll[[2]])
nodes_result = NetzGO(control_mat = netAll[[2]], case_mat = netAll[[1]],
    control_dist = modAll[[2]], case_dist = modAll[[1]],
    control_mod = as.numeric(modAll[[4]]), case_mod = as.numeric(modAll[[3]]))

}
